requirements.txt
agent-framework-core==1.0.0rc3
agent-framework-azure-ai==1.0.0rc3
azure-ai-agentserver-agentframework==1.0.0b16
azure-ai-agentserver-core==1.0.0b16
azure-identity==1.25.1
pyodbc==5.2.0
python-dotenv==1.1.0
